comienzo HUD: mapa, mochila, inventario y boleadoras
