--- Assets/Scripts/Audio.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ec16cc30c1db45a4eafa94b7c770fb08
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
