--- Assets/Scripts/Audio/FMODTester.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8de7dc27cb2d274408acf9ef2f3c1241
