--- Assets/Scripts/Audio/PlayerFootStepAudio.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 80e3548f5e9d2c245ba57f7f267a7e23
