+++ Assets/Scripts/Inventory/InventoryUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: e2665b1f1d003c846a59a295814d7956
